Exception handling

+++ OnlineShopCQRS.Application/Products/Commands/DeleteProduct/DeleteProductCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿
 using MediatR;
+using OnlineShopCQRS.Domain.Exceptions;
 using OnlineShopCQRS.Domain.Repository;
 
 namespace OnlineShopCQRS.Application.Products.Commands.DeleteProduct
@@ -10,12 +11,25 @@ namespace OnlineShopCQRS.Application.Products.Commands.DeleteProduct
 
         public DeleteProductCommandHandler(IProductRepository productRepository)
         {
+
             _productRepository = productRepository ?? throw new ArgumentNullException(nameof(productRepository));
         }
 
         public async Task<int> Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
-            return await _productRepository.DeleteProductAsync(request.Id);
+            var product = await _productRepository.GetProductByIdAsync(request.Id);
+            if (product == null)
+            {
+                throw new NotFoundException($"Produkt ID {request.Id} not exist.");
+            }
+            var deletedCount =  await _productRepository.DeleteProductAsync(request.Id);
+            if (deletedCount == 0)
+            {
+                throw new ProductDeletionException($"Failed to remove product ID {request.Id}.");
+            }
+
+            return deletedCount;
+
         }
     }
 }
